insert action
updated

+++ src/test/java/bean/ActionElements.java
@@ -3,7 +3,7 @@ package bean;
 public class ActionElements {
     public String element;
     public String condition;
-    public int timeout;
+    public long timeout;
     public String inputType;
     public String infoType;
 
@@ -31,11 +31,11 @@ public class ActionElements {
         this.condition = condition;
     }
 
-    public int getTimeout() {
+    public long getTimeout() {
         return timeout;
     }
 
-    public void setTimeout(int timeout) {
+    public void setTimeout(long timeout) {
         this.timeout = timeout;
     }
